feat: Refactor private chats to use GUIDs and fix refresh bug
This commit introduces a major security enhancement by replacing user IDs with non-sequential GUIDs in private chat URLs. This prevents potential user enumeration attacks.

Key changes include:
- Added a `PrivateChatConversations` table to store GUID-based conversation identifiers.
- Updated the backend services and API to handle GUIDs for private chats.
- Modified the `HomeController` to differentiate between private (GUID) and group/channel (long) chat IDs.
- Updated the frontend JavaScript in `Index.cshtml` to correctly parse the new URL format and reload chat history on page refresh.
- Fixed a critical Cross-Site Scripting (XSS) vulnerability in `_ChatMessageBody.cshtml` by removing the usage of `@Html.Raw()`.
- Created a new database migration to apply the schema changes.
- Added `PrivateChatDto.cs` and updated related DTOs and services.
- Corrected various compilation errors that arose during the refactoring process.

+++ Messenger.API/Controllers/FileManagementController.cs
@@ -63,7 +63,7 @@ namespace Messenger.API.Controllers
 
       
         [HttpGet("GetSharedFiles")]
-        public async Task<IActionResult> GetSharedFiles([FromQuery] int chatId, [FromQuery] string groupType)
+        public async Task<IActionResult> GetSharedFiles([FromQuery] long chatId, [FromQuery] string groupType)
         {
             if (chatId <= 0 || string.IsNullOrEmpty(groupType))
                 return BadRequest("Valid Chat ID and Group Type are required.");
@@ -87,7 +87,7 @@ namespace Messenger.API.Controllers
         /// <param name="groupType"></param>
         /// <returns></returns>
         [HttpGet("GetCountSharedFiles")]
-        public async Task<IActionResult> GetCountSharedFiles([FromQuery] int chatId, [FromQuery] string groupType)
+        public async Task<IActionResult> GetCountSharedFiles([FromQuery] long chatId, [FromQuery] string groupType)
         {
             if (chatId <= 0 || string.IsNullOrEmpty(groupType))
                 return BadRequest("Valid Chat ID and Group Type are required.");
